LayoutTests/http/tests/security/basic-auth-module-propagate-expected.txt

127.0.0.1:8000 - didReceiveAuthenticationChallenge - ProtectionSpaceAuthenticationSchemeHTTPBasic - Responding with testUser:testPassword
CONSOLE MESSAGE: Blocked https://127.0.0.1:8443/security/resources/protected-script.php?source=module-propagate&testId=2 from asking for credentials because it is a cross-origin request.
CONSOLE MESSAGE: Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: TypeError: Cross-origin script load denied by Cross-Origin Resource Sharing policy.
CONSOLE MESSAGE: Blocked https://localhost:8443/security/resources/protected-script.php?source=module-propagate&testId=3 from asking for credentials because it is a cross-origin request.
CONSOLE MESSAGE: Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: TypeError: Cross-origin script load denied by Cross-Origin Resource Sharing policy.
CONSOLE MESSAGE: Blocked https://127.0.0.1:8443/security/resources/protected-script.php?source=module-propagate&testId=6 from asking for credentials because it is a cross-origin request.
CONSOLE MESSAGE: Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: TypeError: Cross-origin script load denied by Cross-Origin Resource Sharing policy.
CONSOLE MESSAGE: Blocked https://localhost:8443/security/resources/protected-script.php?source=module-propagate&testId=7 from asking for credentials because it is a cross-origin request.
CONSOLE MESSAGE: Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: TypeError: Cross-origin script load denied by Cross-Origin Resource Sharing policy.
CONSOLE MESSAGE: Blocked https://127.0.0.1:8443/security/resources/protected-script.php?source=module-propagate&testId=10 from asking for credentials because it is a cross-origin request.
CONSOLE MESSAGE: Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: TypeError: Cross-origin script load denied by Cross-Origin Resource Sharing policy.
CONSOLE MESSAGE: Blocked https://localhost:8443/security/resources/protected-script.php?source=module-propagate&testId=11 from asking for credentials because it is a cross-origin request.
CONSOLE MESSAGE: Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: TypeError: Cross-origin script load denied by Cross-Origin Resource Sharing policy.
CONSOLE MESSAGE: Blocked http://127.0.0.1:8000/security/resources/protected-script.php?source=module-propagate&testId=12 from asking for credentials because it is a cross-origin request.
CONSOLE MESSAGE: Blocked http://127.0.0.1:8000/security/resources/protected-script.php?source=module-propagate&testId=13 from asking for credentials because it is a cross-origin request.
CONSOLE MESSAGE: Blocked https://127.0.0.1:8443/security/resources/protected-script.php?source=module-propagate&testId=14 from asking for credentials because it is a cross-origin request.
CONSOLE MESSAGE: Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: TypeError: Cross-origin script load denied by Cross-Origin Resource Sharing policy.
CONSOLE MESSAGE: Blocked https://localhost:8443/security/resources/protected-script.php?source=module-propagate&testId=15 from asking for credentials because it is a cross-origin request.
CONSOLE MESSAGE: Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: TypeError: Cross-origin script load denied by Cross-Origin Resource Sharing policy.
CONSOLE MESSAGE: Blocked https://127.0.0.1:8443/security/resources/protected-script.php?source=module-propagate&testId=18 from asking for credentials because it is a cross-origin request.
CONSOLE MESSAGE: Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: TypeError: Cross-origin script load denied by Cross-Origin Resource Sharing policy.
CONSOLE MESSAGE: Blocked https://localhost:8443/security/resources/protected-script.php?source=module-propagate&testId=19 from asking for credentials because it is a cross-origin request.
CONSOLE MESSAGE: Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin.
CONSOLE MESSAGE: TypeError: Cross-origin script load denied by Cross-Origin Resource Sharing policy.
PASS did load script with origin ./protected-script.php?source=module-propagate&testId=0.

PASS did load script with origin http://127.0.0.1:8000/security/resources/protected-script.php?source=module-propagate&testId=1.

PASS did not load script with origin https://127.0.0.1:8443/security/resources/protected-script.php?source=module-propagate&testId=2.

PASS did not load script with origin https://localhost:8443/security/resources/protected-script.php?source=module-propagate&testId=3.

PASS did load script with origin ./protected-script.php?source=module-propagate&testId=4.

PASS did load script with origin http://127.0.0.1:8000/security/resources/protected-script.php?source=module-propagate&testId=5.

PASS did not load script with origin https://127.0.0.1:8443/security/resources/protected-script.php?source=module-propagate&testId=6.

PASS did not load script with origin https://localhost:8443/security/resources/protected-script.php?source=module-propagate&testId=7.

PASS did load script with origin ./protected-script.php?source=module-propagate&testId=8.

PASS did load script with origin http://127.0.0.1:8000/security/resources/protected-script.php?source=module-propagate&testId=9.

PASS did not load script with origin https://127.0.0.1:8443/security/resources/protected-script.php?source=module-propagate&testId=10.

PASS did not load script with origin https://localhost:8443/security/resources/protected-script.php?source=module-propagate&testId=11.

PASS did not load script with origin ./protected-script.php?source=module-propagate&testId=12.

PASS did not load script with origin http://127.0.0.1:8000/security/resources/protected-script.php?source=module-propagate&testId=13.

PASS did not load script with origin https://127.0.0.1:8443/security/resources/protected-script.php?source=module-propagate&testId=14.

PASS did not load script with origin https://localhost:8443/security/resources/protected-script.php?source=module-propagate&testId=15.

PASS did load script with origin ./protected-script.php?source=module-propagate&testId=16.

PASS did load script with origin http://127.0.0.1:8000/security/resources/protected-script.php?source=module-propagate&testId=17.

PASS did not load script with origin https://127.0.0.1:8443/security/resources/protected-script.php?source=module-propagate&testId=18.

PASS did not load script with origin https://localhost:8443/security/resources/protected-script.php?source=module-propagate&testId=19.

PASS successfullyParsed is true

TEST COMPLETE
